+++ flexmeasures/data/schemas/account.py
@@ -2,6 +2,7 @@ from flask.cli import with_appcontext
 from flexmeasures.data import ma
 from marshmallow import fields
 
+from flexmeasures.data import db
 from flexmeasures.data.models.user import (
     Account as AccountModel,
     AccountRole as AccountRoleModel,
@@ -38,7 +39,7 @@ class AccountIdField(fields.Int, MarshmallowClickMixin):
     @with_appcontext
     def _deserialize(self, value, attr, obj, **kwargs) -> AccountModel:
         """Turn an account id into an Account."""
-        account = AccountModel.query.get(value)
+        account = db.session.get(AccountModel, value)
         if account is None:
             raise FMValidationError(f"No account found with id {value}.")
         # lazy loading now (account somehow is not in the session after this)