+++ flexmeasures/data/schemas/tests/test_sensor.py
@@ -1,5 +1,7 @@
 import pytest
+from flexmeasures import Sensor
 from flexmeasures.data.schemas.sensors import QuantityOrSensor
+from flexmeasures.utils.unit_utils import ur
 from marshmallow import ValidationError
 
 
@@ -55,3 +57,40 @@ def test_quantity_or_sensor_conversion(
 
     schema = QuantityOrSensor(to_unit="MW")
     assert schema.deserialize(src_quantity).magnitude == expected_magnitude
+
+
+@pytest.mark.parametrize(
+    "sensor_id, input_param, dst_unit, fails",
+    [
+        # deserialize a sensor
+        (1, "sensor:1", "MWh", False),
+        (1, "sensor:1", "kWh", False),
+        (1, "sensor:1", "kW", False),
+        (1, "sensor:1", "EUR", True),
+        (2, "sensor:2", "EUR/kWh", False),
+        (2, "sensor:2", "EUR", True),
+        # deserialize a quantity
+        (None, "1MWh", "MWh", False),
+        (None, "1 MWh", "kWh", False),
+        (None, "1 MWh", "kW", True),
+        (None, "100 EUR/MWh", "EUR/kWh", False),
+        (None, "100 EUR/MWh", "EUR", True),
+    ],
+)
+def test_quantity_or_sensor_field(
+    setup_dummy_sensors, sensor_id, input_param, dst_unit, fails, db
+):
+
+    field = QuantityOrSensor(to_unit=dst_unit)
+
+    try:
+        if sensor_id is None:
+            val = field.convert(input_param, None, None)
+            assert val.units == ur.Unit(dst_unit)
+        else:
+            val = field.convert(input_param, None, None)
+            assert val == db.session.get(Sensor, sensor_id)
+
+        assert not fails
+    except Exception:
+        assert fails