+++ flexmeasures/data/services/sensors.py
@@ -9,6 +9,7 @@ from flexmeasures.data.models.time_series import TimedBelief
 
 import sqlalchemy as sa
 
+from flexmeasures.data import db
 from flexmeasures import Sensor, Account
 from flexmeasures.data.models.generic_assets import GenericAsset
 from flexmeasures.data.schemas.reporting import StatusSchema
@@ -27,7 +28,7 @@ def get_sensors(
     :param sensor_id_allowlist:     optionally, allow only sensors whose id is in this list
     :param sensor_name_allowlist:   optionally, allow only sensors whose name is in this list
     """
-    sensor_query = Sensor.query
+    sensor_query = sa.select(Sensor)
     if account is None:
         account_ids = []
     elif isinstance(account, list):
@@ -50,7 +51,7 @@ def get_sensors(
         sensor_query = sensor_query.filter(Sensor.id.in_(sensor_id_allowlist))
     if sensor_name_allowlist:
         sensor_query = sensor_query.filter(Sensor.name.in_(sensor_name_allowlist))
-    return sensor_query.all()
+    return db.session.scalars(sensor_query).all()
 
 
 def get_most_recent_knowledge_time(