+++ documentation/cli/change_log.rst
@@ -4,8 +4,29 @@
 FlexMeasures CLI Changelog
 **********************
 
-since v0.19.0 | February xx, 2024
+since v.0.20.0 | March XX, 2024
+=================================
+
+* Add command ``flexmeasures edit transfer-ownership`` to transfer the ownership of an asset and its children.
+
+since v0.19.1 | February 26, 2024
 =======================================
+
+* Fix support for providing a sensor definition to the ``--storage-power-capacity`` option of the ``flexmeasures add schedule for-storage`` command.
+
+since v0.19.0 | February 18, 2024
+=======================================
+
+* Enable the use of QuantityOrSensor fields for the ``flexmeasures add schedule for-storage`` CLI command:
+
+    * ``charging-efficiency``
+    * ``discharging-efficiency``
+    * ``soc-gain``
+    * ``soc-usage``
+    * ``power-capacity``
+    * ``production-capacity``
+    * ``consumption-capacity``
+    * ``storage-efficiency``
 
 * Streamline CLI option naming by favoring ``--<entity>`` over ``--<entity>-id``. This affects the following options:
 
@@ -16,12 +37,10 @@ since v0.19.0 | February xx, 2024
     * ``--source-id`` -> ``--source``
     * ``--user-id`` -> ``--user`
 
-
-since v0.18.1 | January 15, 2023
+since v0.18.1 | January 15, 2024
 =======================================
 
 * Fix the validation of the option ``--parent-asset`` of command ``flexmeasures add asset``.
-
 
 since v0.17.0 | November 8, 2023
 =======================================